requirements.txt
aiogram
python-chess
cairosvg
